Relocate context to pages/projectSettings dir
