+++ src/pages/projectSettings/Tabs.tsx
@@ -2,8 +2,8 @@ import { useEffect, useMemo, ComponentType } from "react";
 import styled from "@emotion/styled";
 import { Route, useParams } from "react-router-dom";
 import { routes, ProjectSettingsTabRoutes } from "constants/routes";
-import { useProjectSettingsContext } from "context/projectSettings";
 import { ProjectSettingsQuery, RepoSettingsQuery } from "gql/generated/types";
+import { useProjectSettingsContext } from "./Context";
 import { Header } from "./Header";
 import { NavigationModal } from "./NavigationModal";
 import {
