+++ src/pages/projectSettings/tabs/GithubCommitQueueTab/GithubCommitQueueTab.tsx
@@ -6,7 +6,7 @@ import { ProjectSettingsTabRoutes } from "constants/routes";
 import {
   usePopulateForm,
   useProjectSettingsContext,
-} from "context/projectSettings";
+} from "pages/projectSettings/Context";
 import { environmentalVariables } from "utils";
 import { ProjectType } from "../utils";
 import { getFormSchema } from "./getFormSchema";
